fix(ts,connection) avoid mixing web and native code
